Input responsiveness done
